Updating readme and change a function to a separeted file to get better tests
