+++ README.md
@@ -1,1 +1,3 @@
 # ImagesViewer
+
+A aplication to learn how to use table view and segue a new view with a image
