+++ SeeImages/ViewController.swift
@@ -15,17 +15,6 @@ class ViewController: UITableViewController {
     var pictures: [String] = []
 //    var pictures = [String]()
     
-    func orderImg(_ pictures: [String]) -> [String] {
-        guard pictures.count > 1 else { return pictures}
-        
-        let pivot = pictures[pictures.count / 2]
-        let less = pictures.filter {$0 < pivot}
-        let equal = pictures.filter {$0 == pivot}
-        let greater = pictures.filter {$0 > pivot}
-        
-        return orderImg(less) + equal + orderImg(greater)
-    }
-    
     override func viewDidLoad() {
         super.viewDidLoad()
